c2

+++ deleteappli.php
@@ -11,7 +11,7 @@ if(!isset($user_id)){
  
  if (isset($_POST['delete'])) {
     $id = $_POST['id'];
-    mysqli_query($conn, "DELETE FROM `jobapplied` WHERE id1 = '$id'") or die('query failed');
+    mysqli_query($conn, "DELETE FROM `jobapplied` WHERE id1 = '$id' AND uid='$user_id'") or die('query failed');
     header('location:myjob.php');
  }
